major refactor of code
do everything in the session, its easier to keep track of cookies
